# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DayChallenge API",
        "description": "Habit Gamification & Personal Finance API",
        "contact": {
            "name": "DayChallenge Support",
            "email": "support@daychallenge.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Local Server"
        }
    ],
    "paths": {
...
        "/challenges": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "List challenges",
                "description": "Retrieves all available and joined challenges for the user.",
                "operationId": "5f7157215899fabbbf767e5b914a5f47",
                "responses": {
                    "200": {
                        "description": "Challenges retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Challenges retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Challenge"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Create a challenge",
                "description": "Creates a new challenge.",
                "operationId": "b15c3f5c691de3dae50e68ab0ae83b3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreChallengeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Challenge created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Challenge created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Challenge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/challenges/{challenge}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Get challenge details",
                "description": "Retrieves details of a specific challenge.",
                "operationId": "83ff4393c1bcca81606c465d22f53484",
                "parameters": [
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "The ID of the challenge",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Challenge details retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Challenge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Update challenge details",
                "description": "Updates an existing challenge.",
                "operationId": "c1a3d4d672b935238f20782169a1210e",
                "parameters": [
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "The ID of the challenge",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreChallengeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Challenge updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Challenge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Delete a challenge",
                "description": "Deletes a challenge.",
                "operationId": "b4593a03e3e60be380285046501ba1af",
                "parameters": [
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "The ID of the challenge",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Challenge deleted successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Challenge": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9b3d9d30-b6bd-474c-8822-67cc3b922a97"
                    },
                    "title": {
                        "type": "string",
                        "example": "30-Day Savings Challenge"
                    },
                    "description": {
                        "type": "string",
                        "example": "Complete tasks in category Finance to earn bonus coins",
                        "nullable": true
                    },
                    "reward_coins": {
                        "type": "integer",
                        "example": 100
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-30"
                    },
                    "is_joined": {
                        "type": "boolean",
                        "example": true
                    },
                    "participants_count": {
                        "type": "integer",
                        "example": 42
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-09T03:56:49.000000Z"
                    }
                },
                "type": "object"
            },
...
            "StoreChallengeRequest": {
                "required": [
                    "title",
                    "duration_days",
                    "category_ids"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Ultimate Fitness Routine"
                    },
                    "description": {
                        "type": "string",
                        "example": "Achieve streak milestones in Fitness & Health category"
                    },
                    "duration_days": {
                        "type": "integer",
                        "example": 30,
                        "minimum": 1
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": [
                            "9b3d9d30-b4bd-474c-8822-67cc3b922a95"
                        ]
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}